backend-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.5.0
python-dotenv==1.0.0
geopy==2.4.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6